requirements.txt
Flask
Pillow
opencv-python
numpy
matplotlib